pyproject.toml
[tool.poetry]
name = "example_oauth_provider"
version = "0.1.0"
description = "Example OAuth2 Provider in FastAPI"
authors = ["Đỗ Quốc Vương <vuongtlt13@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = {version = "^1.4.42", extras = ["asyncio"]}
inflection = "^0.5.1"
uvicorn = "^0.19.0"
aiomysql = "^0.1.1"
jinja2 = "^3.1.2"
alembic = "^1.8.1"
cryptography = "^38.0.1"
fastapi-sso = "^0.6.3"
python-multipart = "^0.0.5"
fastapi-oauth = "^0.1.3"

[tool.poetry.dev-dependencies]


[[tool.poetry.source]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
default = false
secondary = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
